+++ rolling/__init__.py
@@ -3,6 +3,7 @@ from .arithmetic import Sum, Product, Nunique
 from .entropy import Entropy
 from .hash import PolynomialHash
 from .logical import All, Any
+from .matching import Match
 from .minmax import Min, Max, MinHeap
 from .stats import Mean, Var, Std, Median, Mode, Skew, Kurtosis